Nightly sync.
I added a start up menu! It has cycling/selection noises, and none of
the options work besides the Quit Game option. Other than that, I plan
on reprogramming warps to make them easier to create and also working on
fading the screen in and out in-between room transitions in order to
recreate that Earthbound-y feel.