Added an executable and changed the menu object.
Changed the menu object so that people who want to test what's in the
game (barely anything) can do it if they so desire. Selecting New Game
will just take you to the first room I have setup. Enter is to select
it. That's about it, enjoy if you want. I'll try and update the
executable every once in a while.