Added fully functional name input!
I have finally added a fully functional keyboard with name input. This
name can be used anywhere, and different names can generate different
messages on the confirm screen (not yet finished.) That's about it, but
I am also taking away the Earthbound part of the game, as I want to keep
the Undertale style and not go to that advanced pixel art style.