Working on a save object and fixed the closing sprites.
Closing now works, and it even include that authentic fade-in effect
from Undertale! I am also starting to work on a save object, but I have
no compiled a new executable and do not plan on doing so until the
saving system is done. I don't want to put up something so buggy.